Fix HOME being changed
Context: there is desire to use pkgs.npmHooks.npmConfigHook in a nix
shell

This fixes* HOME being set to another value and messing up your
environment.

It should be noted that the npm-insall-hook seems to rely on the
variable being changed in the npm-config-hook, as without changing it
back in the npm-install-hook, typescript (and probably many more
packages) failed to build.

Q: What about the exits? You don't set HOME back on failure.
A: `exit` will quit out of the nix shell, so we don't need to worry
about it

*The change I made to npm-install-hook was just the first thing I tried,
and it worked for my project, but it may be a bad solution, and break
other things.